Move Storage interface in interfaces.go
